Reorganize the store Fix AOT errors
