+++ client/app/app.module.ts
@@ -4,10 +4,8 @@ import { HttpModule } from '@angular/http';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { compose } from '@ngrx/core';
 import { EffectsModule } from '@ngrx/effects';
-import { combineReducers, StoreModule } from '@ngrx/store';
-import { storeLogger } from 'ngrx-store-logger';
+import { StoreModule } from '@ngrx/store';
 
 import { AppComponent } from './app.component';
 import { AuthenticationService } from './authentication.service';
@@ -15,16 +13,8 @@ import { LoginModalComponent } from './login-modal/login-modal.component';
 import { NavigationComponent } from './navigation/navigation.component';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 import { AppRoutes, AuthenticationGuard } from './routes';
-import { DEFAULT_APPLICATION_STATE } from './store/application-state';
 import { AuthenticationEffectService } from './store/effects/authentication-effect.service';
-import { session } from './store/session-reducer';
-
-export function rootReducer(state: any, action: any) {
-  return compose(
-    storeLogger({ collapsed: true }),
-    combineReducers
-  )({ session })(state, action);
-}
+import { rootReducer } from './store/reducers/root-reducer';
 
 @NgModule({
   declarations: [
@@ -46,7 +36,7 @@ export function rootReducer(state: any, action: any) {
     ReactiveFormsModule,
     NgbModule.forRoot(),
     RouterModule.forRoot(AppRoutes),
-    StoreModule.provideStore(rootReducer, DEFAULT_APPLICATION_STATE),
+    StoreModule.provideStore(rootReducer),
     EffectsModule.run(AuthenticationEffectService)
   ],
   bootstrap: [AppComponent]
