--- client/app/store/application-state.ts
@@ -1,9 +1,0 @@
-import { ISessionState, SessionState } from './session-reducer';
-
-export class IApplicationState {
-  session: ISessionState;
-}
-
-export const DEFAULT_APPLICATION_STATE: IApplicationState = {
-  session: new SessionState()
-};
