+++ client/app/store/records/application-state.ts
@@ -1,0 +1,5 @@
+import { ISessionState } from './session-state';
+
+export class IApplicationState {
+  session: ISessionState;
+}
